// src/app/settings/models/settings-custom-lable-model.ts
export class SettingsCustomLableModel {
    id: number
    institution_id: number;
    batch: string;
    batch_plural: string;
    program: string;
    program_plural: string;
    course: string;
    course_plural: string;
    normal_subject: string;
    normal_subject_plural: string;
    elective_subject: string;
    elective_subject_plural: string;
    elective_group: string;
    elective_group_plural: string;
    subject_group: string;
    subject_group_plural: string;
    module_chapter: string;
    module_chapter_plural: string;
    topics_skills: string;
    topics_skills_plural: string;
    is_active: boolean;
}